Data from the Experiment B006

These data are from an experiment on touch perception in humans that applied the so-called Limb Identification Task. In each trial, participants receive two stimuli that were either on the same arm or on two different arms and decided by lifting either their heels or toes whether they were on the same or on different limbs. Of interest is a limb crossing effect in this task, i.e., when the limbs are crossed, participants make mistakes and respond slowly despite the fact that the task does not require participants to take posture or spatial aspects into account in their decisions. For details see the publication under https://doi.org/10.1101/2024.02.01.578364.

Usage

B006Data

Format

A data frame with 32455 rows and 30 variables:

Participant

Participant identifier

Run

Experiment run (e.g., 'Run1', 'Run2')

Posture

Posture of the trial (e.g., 'uncrossed')

StimSide

Side of the stimulus (e.g., 'same', 'different')

StimSegment

Segment of the arm where the stimulus was applied

Tactor1

First tactor number

Tactor2

Second tactor number

Tactor1Side

Side of the first tactor (e.g., 'left', 'right')

Tactor1Segment

Segment of the arm for the first tactor

Tactor2Side

Side of the second tactor (e.g., 'left', 'right')

Tactor2Segment

Segment of the arm for the second tactor

Stim1Intens

Intensity of the first stimulus

Stim2Intens

Intensity of the second stimulus

Trialtype

Type of the trial

SOA

Stimulus Onset Asynchrony (interval between stimuli)

Repetition

Repetition count

Trialnumber

Number of the trial

TimeAfterBeep

Time after the beep (timing indicator)

Response

Participant's response

Response_meaning

Meaning of the response

Response_correct

Whether the response was correct

RT

Reaction time in milliseconds

Order

Order of the trial

ResponseCoding

Coding of the response

Trial_ok

Whether the trial was valid

Trial_repeated

Whether the trial was repeated

Age

Age of the participant

Sex

Sex of the participant (e.g., 'male', 'female', 'other')

Arm_top

The arm that was on top in the crossed condition

Length_lower_arm

Length of the participant's lower arm in centimeters

Source

Data collected in a behavioral neuroscience experiment at Tobias Heed's Reach and Touch Lab.