Validate and preprocess the input data for BIDS formatting

This function checks that certain requirements are met by the dataset, including variable names in snake_case, valid participant IDs, and session description. The function can optionally also check for proper gender encoding ("m", "f", or "o"). It attempts to convert invalid data and outputs warnings or errors if issues arise.

Usage

check_input_data(
  data,
  participant_col = "participant_id",
  session_col = "session",
  gender_col = NULL
)

Arguments

data

A data frame that contains the dataset to be validated.

participant_col

Name of the column containing participant IDs (default is "participant_id").

session_col

Name of the column containing session IDs (default is "session").

gender_col

Name of the column containing gender information, if any (default is NULL, assuming there is no gender column).

Value

A validated and possibly modified data frame with the appropriate formatting.

Examples

# Example dataset
example_data <- data.frame(ParticipantID = c(1, 2), Session = c(1, 1), Gender = c("Male", "Female"))
check_input_data(data = example_data, participant_col = "ParticipantID", session_col = "Session", gender_col = "Gender")
#> 
#> Step 1: checking variable labels
#> checking variable Gender
#> renamed variable label: Male -> male
#> renamed variable label: Female -> female
#> 
#> Step 2: checking participant and session identifiers
#> renamed variable: ParticipantID -> participant_id
#> renamed ids to sub-001, sub-002 ...
#> renamed variable: Session -> session
#> renamed sessions to ses-01 ...
#> renamed variable: Gender -> gender
#> 
#> Step 3: checking if all variable names are snake_case
#> renamed variable: NA -> NA
#> renamed variable:  -> 
#> 
#> Step 4: checking correct coding of gender variable
#> Warning: Invalid values found in gender column. Please recode to 'm', 'f', or 'o'.
#>   participant_id session gender
#> 1        sub-001  ses-01   male
#> 2        sub-002  ses-01 female