Read BIDS-formatted (behavioral) data

This function searches the BIDS directory for all task-related .tsv files and reads the data into R. It handles cases where session folders are present as well as cases where there are no session folders.

Usage

read_bids(
  bids_dir,
  data_type = "beh",
  filename_variables = c("sub-", "ses-"),
  variable_names = c("participant", "session"),
  strip_prefixes = F
)

Arguments

bids_dir

The root directory of the BIDS data set.

data_type

A string defining the data_type for the task files. In BIDS it is usually "events" or "beh" (the default) for "behavior".

filename_variables

A character string defining the variables that are contained in the data file name and that should then appear as columns in the data. Default is c("sub-", "ses-") assuming that the file name will at least contain the participant code and the session.

variable_names

A character string defining how the variables defined in 'filename_variables' should be called in the data Default is c("participant", "session").

strip_prefixes

Defines whether or not prefixes like "sub-" should be stripped or maintained. Default is FALSE which would maintain the prefixes.

Value

A list containing the participant data and the task data.

Examples

data <- data.frame(participant_id = c("sub-1", "sub-1", "sub-2", "sub-2"),
                   session = c("ses-1", "ses-2", "ses-1", "ses-2"),
                   run = c("run-1", "run-1", "run-1", "run-1"),
                   age = c(25, 25, 30, 30),
                   sex = c('m', 'm', 'f', 'f'),
                   response_time = c(100, 200, 150, 180))

write_task_tsv(data, bids_dir = "BIDS", filename_prefixes = c("sub-", "ses-", "task-", "run-"),
               filename_variables = c("participant_id", "session", "RTTask"="task", "run"), ignore_variables =  c("age", "sex"))
#> Variable task does not exist in the data and will be imputed as RTTask
#> Main BIDS directory successfully created: BIDS
#> Folder successfully created: BIDS/sub-1/ses-1
#> Task data saved: BIDS/sub-1/ses-1/sub-1_ses-1_task-RTTask_run-1_beh.tsv
#> Folder successfully created: BIDS/sub-1/ses-2
#> Task data saved: BIDS/sub-1/ses-2/sub-1_ses-2_task-RTTask_run-1_beh.tsv
#> Folder successfully created: BIDS/sub-2/ses-1
#> Task data saved: BIDS/sub-2/ses-1/sub-2_ses-1_task-RTTask_run-1_beh.tsv
#> Folder successfully created: BIDS/sub-2/ses-2
#> Task data saved: BIDS/sub-2/ses-2/sub-2_ses-2_task-RTTask_run-1_beh.tsv
write_participants_tsv(data, bids_dir = "BIDS", include_variables = c("participant_id", "age", "sex"))
#> Participants data saved: BIDS/participants.tsv
read_bids(bids_dir = "BIDS")
#> Task data loaded from all task files.
#> Participants data loaded from: BIDS/participants.tsv
#> $participants
#>   participant_id age sex
#> 1          sub-1  25   m
#> 2          sub-2  30   f
#> 
#> $task_data
#>   participant session response_time
#> 1       sub-1   ses-1           100
#> 2       sub-1   ses-2           200
#> 3       sub-2   ses-1           150
#> 4       sub-2   ses-2           180
#>