Create BIDS-compliant behavioral data files

This function creates BIDS-compliant (behavioral) data files in .tsv format. It generates individual task-based files for each participant and session.

Usage

write_task_tsv(
  data,
  bids_dir,
  data_type = "beh",
  filename_prefixes = c("sub-", "ses-", "task-", "run-"),
  filename_variables = c("participant_id", "session", "task", "run"),
  path_prefixes = c("sub-", "ses-"),
  path_variables = c("participant_id", "session"),
  ignore_variables = c(),
  include_variables = c()
)

Arguments

data

A data frame containing the participant and other information and (behavioral) data.

bids_dir

The directory where the BIDS data set will be saved.

data_type

A string defining the data_type for the task files. In BIDS it is usually "events" or "beh" (the default) for "behavior"

filename_prefixes

Character vector containing the prefixes used for the variables appearing in the file names of the tsv files. Default is c("sub-", "ses-", "task-", "run-") assuming that the file names would contain information about the partcipant, session, the task, and the run (in that order). Variables can be added or removed ad lib as long as the vector matches the information provided in the filename_variables vector.

filename_variables

Character vector containing names of the variables in the data set that provide the information added to the filename_prefixes. Default is c("participant_id", "session", "task", "run"). Note that the lengths of filename_prefixes and filename_variables need to match. filename_variables can also define information that is NOT in the data set. In this case, a named vector can be provided. You can for example use c("participant_id", "session", "RTTask"="task", "run") if there is no variable "task" in the data set and then the task will be called "RTTask".

path_prefixes

Character vector containing the prefixes for the filepath structure. Default is c("sub-", "ses-") resulting in a folder structure with subject/participants on the top level and folders for the sessions on the secondary level.

path_variables

Analogous to filename_variables this is the character vector containing names of the variables used for path_prefixes. Default is c("participant_id", "session").

ignore_variables=c()

A character vector of variables to explicitly exclude from the tsv file. If data contains participant information variables, they can be defined here as a character vector, for example as ignore_variables = c("age", "sex").

include_variables=c()

A character vector of variables to explicitly include in the tsv file. This can be useful in case you want to keep one of the variables that end up in the file name in the data.

Value

This function does not return anything but writes files to the output directory.

Examples

example_task_data <- data.frame(participant_id = c("sub-1", "sub-1", "sub-2", "sub-2"),
                   session = c("ses-1", "ses-2", "ses-1", "ses-2"),
                   run = c("run-1", "run-1", "run-1", "run-1"),
                   response_time = c(100, 200, 150, 180))

write_task_tsv(example_task_data2, bids_dir = "example_bids", filename_prefixes = c("sub-", "ses-", "task-", "run-"),
               filename_variables = c("participant_id", "session", "RTTask"="task", "run"), ignore_variables =  c("age", "sex"))
#> Error in eval(expr, envir, enclos): object 'example_task_data2' not found
#clean up
unlink("example_bids/", recursive = T)